# api/api.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Hacker news - HBN13
  description: >-
    This is the Open API documentation for the REST API of our beloved
    application **Hacker news** deployed at
    <https://asw-api-demo.herokuapp.com/>. <br>All operations are executable.
    Only one operation requires authentication: `deleteComment`. In this case,
    you must **Authorize** your request by providing the api_key vaule you got
    when you created the comment.
servers:
  - url: https://serene-ridge-36448.herokuapp.com/API/v1.0/
paths:
...
  /comments:
    post:
      tags:
        - comments
      summary: Afegeix un nou Comentari a una Submission de Hacker News
      requestBody:
        description: Proporcioneu el contingut del nou Comentari
        content:
          application/json:
            schema:
              required:
                - comment
                - id_submission
              type: object
              properties:
                id_submission:
                  type: integer
                  example: 1
                comment:
                  type: string
                  minLength: 4
                  maxLength: 280
                  example: My comment
                id_comment_father:
                  type: integer
                  example: 1
        required: true
      responses:
        '201':
          description: Comentari creat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddCommentResult'
        '400':
          description: 'Error: Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
                400_insuficient_parameters:
                  $ref: '#/components/examples/error400_insuficient_parameters'
                400SubmissionDeleted:
                  value: |-
                    {
                        "status": 400,
                        "error": "Bad Request",
                        "message": "Submission with id: 1 was deleted before."
                    }
        '401':
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        '404':
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
      security:
        - ApiKeyAuth: []
...
  /comment/{id}/edit:
    put:
      tags:
        - comments
      summary: Editeu un comentari donat el seu identificador
      parameters:
        - name: id
          in: path
          description: ID del commentari a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Proporcioneu el contingut del nou comentari
        content:
          application/json:
            schema:
              required:
                - comment
              type: object
              properties:
                comment:
                  type: string
                  minLength: 4
                  maxLength: 280
                  example: My new comment
        required: true
      responses:
        '203':
          description: Comment edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddCommentResult'
        '400':
          description: 'Error: Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        '401':
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                Unauthorized:
                  value: |-
                    {
                        "status": 401,
                        "error": "Unauthorized",
                        "message": "API key not found"
                    }
        '403':
          description: 'Error: Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                IncorrectApiKey:
                  value: |-
                    {
                        "status": 403,
                        "error": "Forbidden",
                        "message": "Incorrect apiKey"
                    }
        '404':
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                NotFound:
                  value: |-
                    {
                        "status": 404,
                        "error": "Not Found",
                        "message": "Comment with id: 123 doesn't exist in our database"
                    }
      security:
        - ApiKeyAuth: []
...
components:
  schemas:
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        author:
          type: string
          example: Electrike
        comment:
          type: string
          minLength: 5
          maxLength: 256
          example: Ejemplo de comentario.
        upVotes:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        id_submission:
          type: integer
          format: int64
          example: 1
        id_comment_father:
          type: integer
          format: int64
          example: 2
        comments:
          type: object
...
    AddCommentResult:
      type: object
      properties:
        status:
          type: integer
          enum:
            - 201
        message:
          type: string
          enum:
            - 'Comment with id: [id] was successfully created.'
        comment:
          $ref: '#/components/schemas/Comment'
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum:
            - 400
            - 401
            - 403
            - 404
            - 409
        error:
          type: string
          enum:
            - Bad Request
            - Unauthorized
            - Forbidden
            - Not Found
            - Conflict
        message:
          type: string
...
  examples:
    error400:
      value:
        status: 400
        error: "Bad Request"
        message: "Content is too long"
    error400_insuficient_parameters:
      value:
        status: 400
        error: Bad Request
        message: Insuficient parameters, missing id_submission
    error401:
      value:
        status: 401
        error: "Unauthorized"
        message: "You didn't provide an api key (X-API-KEY Header)"
...
    error404:
      value:
        status: 404
        error: "Not Found"
        message: "No the ID doesn't exist"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-KEY
      in: header